Release/1.3.0 (#24)
* Move to ghcr.io/gp-devenv/gp-node-devenv, remove Angular and Chromium, improve script, README.md

* Move to ghcr.io/gp-devenv/gp-node-devenv, remove Angular and Chromium, improve script, README.md

* Add cleanup workflow

* Fix workflows

* Remove angular cli

* Remove image dependency updates

* Remove image dependency updates

---------

+++ src/scripts/dev/container/rm.sh
@@ -15,10 +15,20 @@
 
 set -e
 
+if [ -z "$1" ]; then
+    echo "Usage: $0 <UBUNUT_VERSION> <NODE_VERSION>"
+    exit 1
+fi
+
+if [ -z "$2" ]; then
+    echo "Usage: $0 <UBUNUT_VERSION> <NODE_VERSION>"
+    exit 1
+fi
+
 VERSION=`cat .version`-dev
 IMAGE_NAME="`cat .image_name`"
-IMAGE="$IMAGE_NAME:$1-$VERSION"
-CONTAINER="`cat .image_name | sed -e 's/ghcr.io\///g' -e 's/gpfister\///g'`-$1-$VERSION"
+IMAGE="$IMAGE_NAME:$1-$2-$VERSION"
+CONTAINER="`cat .image_name | sed -e 's/ghcr.io\///g' -e 's/gp-devenv\///g'`-$1-$2-$VERSION"
 
 docker container rm $CONTAINER
